FEM: elmer, add experimental report framework
The report framework will be used to report all errors resulting from
executing the solver (RunSolver command). Reports in the report view are
easy to miss and there is all kind of other stuff there. For this reason
the new report framework reports via a QDialog if the gui is loaded.